Large commit for Rivet integration
This commit imports the [O2Rivet](https://gitlab.com/cholmcc/O2Rivet)
project into O2Physics.

This project aims to enable running Rivet analyses on existing
and future simulation productions, locally, on CVMFS, on Grid,
and with HyperLoop.

The main component is the DPL `o2-analysis-mm-rivet`.  This reads in MC
AODs (via f.ex. `o2-aod-mc-producer-workflow`), converts the events to
HepMC events, and then runs Rivet analyses on these.  The output is an
`o2::rivet::RivetAOs` object stored in the generated
`AnalysisResults.root` file.

The tool supports partial runs (e.g., parallel processing on Grid), and
merging, including the `Rivet::Analysis::finalize` step, via regular
ROOT merging.

The tool also supports compiling analyses on the fly.  This is useful in
Grid or similar settings, and for developing and testing analyses not
yet accepted into an official repository.

Examples of using the tool are given in `PWGMM/Rivet/doc` along with
some other documentation.

Compilation of the tool is contingent on the presence of Rivet (and its
requirements YODA, FastJetContrib, and HepMC3) on the target system.
The feature is marked as _recommended_ via optional, but recommended,
build dependencies on Rivet and YODA.